move subrecord picker to utils
